all: Do not store the chain head block in SubgraphDeployment
Storing it there requires that we update every subgraph deployment for a
network whenever we get a new block for that network which causes a huge
amount of writes to subgraphs.subgraph_deployment, and bloats the table.

The information is already available in ethereum_networks; instead of
denormalizing that data, make it available to GraphQL queries through a
view in the database. The view pretends to be an entity, though any attempt
to write that entity will fail as it is not possible to write into this
view.